created initial views for prescriptions and appts.
created a function to be used in conjunction with
user_passes_test decorator to check if the user is
a doctor or nurse.

Each view checks to see if the user is a doctor or
nurse, before returning a dictionary that contains
all the appointments/prescriptions associated with
the user.